refactor: Menu 페이지를 common 폴더 내부로 위치 변경
